requirements.txt

pyyaml>=6.0
python-dotenv>=1.0.0
crawl4ai==0.7.4
pydantic>=1.10,<3

pytest

requests
beautifulsoup4>=4.12
playwright>=1.35
